Refactored XxeAnalyzerTest. Split out other classes into separate files, mad var declarations more consistent, removed comments which appear to not be useful, added TestCategory tags and made Expected vars a field to remove duplication
