+++ SecurityCodeScan.Test/Tests/XxeAnalyzerTest.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
@@ -7,99 +6,12 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.Web.XmlTransform;
 using SecurityCodeScan.Analyzers;
-using SecurityCodeScan.Analyzers.Taint;
 using SecurityCodeScan.Test.Helpers;
-using DiagnosticVerifier = SecurityCodeScan.Test.Helpers.DiagnosticVerifier;
 
 namespace SecurityCodeScan.Test.XXE
 {
-    public abstract class BaseXxeAnalyzerTest : DiagnosticVerifier
-    {
-        private static readonly Version SafeVersion       = new Version(4, 5, 2);
-        private static readonly Version VulnerableVersion = new Version(4, 5, 1);
-
-        protected async Task VerifyDiagnosticSafeUnsafe(string cSharpTest, string visualBasicTest, DiagnosticResult[] expected)
-        {
-            if (cSharpTest == null && visualBasicTest == null)
-                throw new ArgumentException();
-
-            if (cSharpTest != null)
-            {
-                await VerifyCSharpDiagnostic(cSharpTest, expected, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-                // defaults are safe starting 4.5.2
-                await VerifyCSharpDiagnostic(cSharpTest, dotNetVersion: SafeVersion).ConfigureAwait(false);
-            }
-
-            if (visualBasicTest != null)
-            {
-                await VerifyVisualBasicDiagnostic(visualBasicTest, expected, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-                // defaults are safe starting 4.5.2
-                await VerifyVisualBasicDiagnostic(visualBasicTest, dotNetVersion: SafeVersion).ConfigureAwait(false);
-            }
-        }
-
-        protected async Task VerifyAlwaysWarnigs(string cSharpTest, string visualBasicTest, DiagnosticResult[] expected)
-        {
-            if (cSharpTest == null && visualBasicTest == null)
-                throw new ArgumentException();
-
-            if (cSharpTest != null)
-            {
-                await VerifyCSharpDiagnostic(cSharpTest, expected, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-                await VerifyCSharpDiagnostic(cSharpTest, expected, dotNetVersion: SafeVersion).ConfigureAwait(false);
-            }
-
-            if (visualBasicTest != null)
-            {
-                await VerifyVisualBasicDiagnostic(visualBasicTest, expected, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-                await VerifyVisualBasicDiagnostic(visualBasicTest, expected, dotNetVersion: SafeVersion).ConfigureAwait(false);
-            }
-        }
-
-        protected async Task VerifyNoWarnings(string cSharpTest, string visualBasicTest)
-        {
-            if (cSharpTest == null && visualBasicTest == null)
-                throw new ArgumentException();
-
-            if (cSharpTest != null)
-            {
-                await VerifyCSharpDiagnostic(cSharpTest, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-                await VerifyCSharpDiagnostic(cSharpTest, dotNetVersion: SafeVersion).ConfigureAwait(false);
-            }
-
-            if (visualBasicTest != null)
-            {
-                await VerifyVisualBasicDiagnostic(visualBasicTest, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-                await VerifyVisualBasicDiagnostic(visualBasicTest, dotNetVersion: SafeVersion).ConfigureAwait(false);
-            }
-        }
-
-        public enum Warnings
-        {
-            None,
-            Always,
-            OnFramework,
-        }
-
-        protected async Task Verify(string cSharpTest, string visualBasicTest, DiagnosticResult[] expected, Warnings expectWarnings)
-        {
-            switch (expectWarnings)
-            {
-                case Warnings.Always:
-                    await VerifyAlwaysWarnigs(cSharpTest, visualBasicTest, expected).ConfigureAwait(false);
-                    break;
-                case Warnings.OnFramework:
-                    await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, expected).ConfigureAwait(false);
-                    break;
-                case Warnings.None:
-                    await VerifyNoWarnings(cSharpTest, visualBasicTest).ConfigureAwait(false);
-                    break;
-            }
-        }
-    }
-
     [TestClass]
-    public class XxeAnalyzerTest : BaseXxeAnalyzerTest
+    public class XxeAnalyzerTest : XxeAnalyzerTestBase
     {
         protected override IEnumerable<DiagnosticAnalyzer> GetDiagnosticAnalyzers(string language)
         {
@@ -111,16 +23,18 @@ namespace SecurityCodeScan.Test.XXE
             MetadataReference.CreateFromFile(typeof(XmlFileInfoDocument).Assembly.Location),
         };
 
+        /// <summary> XML parsing vulnerable to XXE </summary>
+        private DiagnosticResult[] Expected = new[]
+        {
+            new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
+        };
+
         protected override IEnumerable<MetadataReference> GetAdditionalReferences() => References;
 
+        [TestCategory("Detect")]
         [TestMethod]
         public async Task XPathDocument()
         {
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
             const string cSharpTest = @"
 using System.Xml.XPath;
 
@@ -133,7 +47,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest = @"
+            const string visualBasicTest = @"
 Imports System.Xml.XPath
 
 Class Xxe
@@ -144,7 +58,7 @@ Class Xxe
 End Class
 ";
 
-            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, Expected).ConfigureAwait(false);
 
             const string cSharpTest2 = @"
 using System.Xml.XPath;
@@ -157,7 +71,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest2 = @"
+            const string visualBasicTest2 = @"
 Imports System.Xml.XPath
 
 Class Xxe
@@ -167,22 +81,18 @@ Class Xxe
 End Class
 ";
 
-            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, Expected).ConfigureAwait(false);
         }
 
+        [TestCategory("Detect")]
         [DataRow("new StringReader(path)", Warnings.OnFramework)]
         [DataRow("XmlReader.Create (path, new XmlReaderSettings (){DtdProcessing = DtdProcessing.Parse})", Warnings.OnFramework)]
         [DataTestMethod]
         public async Task XmlSchemaRead(string createReader, Warnings expectWarnings)
         {
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
             string vbCreateReader = createReader.Replace("(){", "() With {.");
 
-            string cSharpTest = $@"
+            var cSharpTest = $@"
 #pragma warning disable 8019
     using System.Xml.Schema;
     using System.IO;
@@ -212,7 +122,7 @@ Class Xxe
     End Sub
 End Class
 ";
-            await Verify(cSharpTest, visualBasicTest, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest, visualBasicTest, Expected, expectWarnings).ConfigureAwait(false);
 
             string cSharpTest2 = $@"
 #pragma warning disable 8019
@@ -242,9 +152,10 @@ Class Xxe
     End Sub
 End Class
 ";
-            await Verify(cSharpTest2, visualBasicTest2, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest2, visualBasicTest2, Expected, expectWarnings).ConfigureAwait(false);
         }
 
+        [TestCategory("Detect")]
         [DataRow("XmlDocument",        "Load(text)",      Warnings.OnFramework)]
         [DataRow("XmlDocument",        "LoadXml(text)",   Warnings.OnFramework)]
         [DataRow("XmlDocument",        "InnerXml = text", Warnings.OnFramework)]
@@ -258,12 +169,7 @@ End Class
         [DataTestMethod]
         public async Task XmlDocumentDefaultsClassMember(string className, string sink, Warnings expectWarnings)
         {
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
-            string cSharpTest3 = $@"
+            var cSharpTest3 = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -283,7 +189,7 @@ class Xxe
     }}
 }}";
 
-            string visualBasicTest3 = !sink.Contains("=") ? $@"
+            var visualBasicTest3 = !sink.Contains("=") ? $@"
 #Disable Warning BC50001
     Imports System.Xml
     Imports System.Configuration
@@ -304,25 +210,10 @@ Class Xxe
 End Class
 " : null; // todo: how to set property of a temporary object in VB?
 
-            await Verify(cSharpTest3, visualBasicTest3, expected, expectWarnings).ConfigureAwait(false);
-
-            // todo: different VB specific field declaration syntax is not recognized
-            //            var visualBasicTest4 = @"
-            //Imports System.Xml
-
-            //Class Xxe
-            //    Private Shared Parser As New XmlDocument()
-
-            //    Public Shared Sub parseUpload(path As String)
-            //        Parser.Load(path)
-            //    End Sub
-            //End Class
-            //";
-            //            await VerifyVisualBasicDiagnostic(visualBasicTest4, expected, dotNetVersion: VulnerableVersion).ConfigureAwait(false);
-            //            // defaults are safe starting 4.5.2
-            //            await VerifyVisualBasicDiagnostic(visualBasicTest4, dotNetVersion: SafeVersion).ConfigureAwait(false);
+            await Verify(cSharpTest3, visualBasicTest3, Expected, expectWarnings).ConfigureAwait(false);
         }
 
+        [TestCategory("Detect")]
         [DataRow("XmlDocument",         "Load(text)",      Warnings.OnFramework)]
         [DataRow("XmlDocument",         "LoadXml(text)",   Warnings.OnFramework)]
         [DataRow("XmlDocument",         "InnerXml = text", Warnings.OnFramework)]
@@ -340,12 +231,7 @@ End Class
         [DataTestMethod]
         public async Task XmlDocumentDefaults(string className, string sink, Warnings expectWarnings)
         {
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
-            string cSharpTest = $@"
+            var cSharpTest = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -386,9 +272,9 @@ Class Xxe
 End Class
 ";
 
-            await Verify(cSharpTest, visualBasicTest, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest, visualBasicTest, Expected, expectWarnings).ConfigureAwait(false);
 
-            string cSharpTest2 = $@"
+            var cSharpTest2 = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -407,7 +293,7 @@ class Xxe
     }}
 }}";
 
-            string visualBasicTest2 = !sink.Contains("=") ? $@"
+            var visualBasicTest2 = !sink.Contains("=") ? $@"
 #Disable Warning BC50001
     Imports System.Xml
     Imports System.Configuration
@@ -427,7 +313,7 @@ Class Xxe
 End Class
 " : null; // todo: how to set property of a temporary object in VB?
 
-            await Verify(cSharpTest2, visualBasicTest2, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest2, visualBasicTest2, Expected, expectWarnings).ConfigureAwait(false);
 
 // todo: different VB specific field declaration syntax is not recognized
 //            var visualBasicTest4 = @"
@@ -448,6 +334,7 @@ End Class
 
         private const string SecureResolverText = "new XmlSecureResolver(new XmlUrlResolver(), \"http://myLocalSite/\")";
 
+        [TestCategory("Detect")]
         [DataRow("XmlDocument",       "()",     "XmlResolver", "null",                 "Load(text)",      Warnings.None)]
         [DataRow("XmlDocument",       "()",     "XmlResolver", SecureResolverText,     "Load(text)",      Warnings.None)]
         [DataRow("XmlDocument",       "()",     "XmlResolver", "new XmlUrlResolver()", "Load(text)",      Warnings.Always)]
@@ -482,12 +369,7 @@ End Class
             var vbValue = value.Replace("null", "Nothing");
             vbValue = Regex.Replace(vbValue, @"^\(([^\)]*)\)(.*)", "DirectCast($2, $1)", RegexOptions.Multiline);
 
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
-            string cSharpTest = $@"
+            var cSharpTest = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -524,9 +406,9 @@ Class Xxe
 End Class
 ";
 
-            await Verify(cSharpTest, visualBasicTest, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest, visualBasicTest, Expected, expectWarnings).ConfigureAwait(false);
 
-            string cSharpTest2 = $@"
+            var cSharpTest2 = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -560,9 +442,9 @@ Class Xxe
     End Sub
 End Class
 ";
-            await Verify(cSharpTest2, visualBasicTest2, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest2, visualBasicTest2, Expected, expectWarnings).ConfigureAwait(false);
 
-            string cSharpTest3 = $@"
+            var cSharpTest3 = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -579,7 +461,7 @@ class Xxe
     }}
 }}";
 
-            string visualBasicTest3 = !sink.Contains("=") ? $@"
+            var visualBasicTest3 = !sink.Contains("=") ? $@"
 #Disable Warning BC50001
     Imports System.Xml
     Imports System.Configuration
@@ -595,9 +477,9 @@ Class Xxe
 End Class
 "
                                        : null; // todo: how to set property of a temporary object in VB?
-            await Verify(cSharpTest3, visualBasicTest3, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest3, visualBasicTest3, Expected, expectWarnings).ConfigureAwait(false);
 
-            string cSharpTest4 = $@"
+            var cSharpTest4 = $@"
 #pragma warning disable 8019
     using System.Xml;
     using System.Configuration;
@@ -631,17 +513,13 @@ Class Xxe
     End Sub
 End Class
 ";
-            await Verify(cSharpTest4, visualBasicTest4, expected, expectWarnings).ConfigureAwait(false);
+            await Verify(cSharpTest4, visualBasicTest4, Expected, expectWarnings).ConfigureAwait(false);
         }
 
+        [TestCategory("Detect")]
         [TestMethod]
         public async Task XmlTextReaderFile()
         {
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
             const string cSharpTest = @"
 using System.Xml;
 
@@ -654,7 +532,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest = @"
+            const string visualBasicTest = @"
 Imports System.Xml
 
 Class Xxe
@@ -664,7 +542,7 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, Expected).ConfigureAwait(false);
 
             const string cSharpTest2 = @"
 using System.Xml;
@@ -677,7 +555,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest2 = @"
+            const string visualBasicTest2 = @"
 Imports System.Xml
 
 Class Xxe
@@ -686,9 +564,10 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, Expected).ConfigureAwait(false);
         }
 
+        [TestCategory("Ignore")]
         [TestMethod]
         public async Task XmlTextReaderDerived()
         {
@@ -709,7 +588,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest = @"
+            const string visualBasicTest = @"
 Imports System.Xml
 
 Class DerivedXmlTextReader
@@ -730,14 +609,10 @@ End Class
             await VerifyNoWarnings(cSharpTest, visualBasicTest).ConfigureAwait(false);
         }
 
+        [TestCategory("Detect")]
         [TestMethod]
         public async Task XmlTextReaderString()
         {
-            var expected = new[]
-            {
-                new DiagnosticResult { Id = "SCS0007", Severity = DiagnosticSeverity.Warning }
-            };
-
             const string cSharpTest = @"
 using System.Xml;
 using System.IO;
@@ -751,7 +626,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest = @"
+            const string visualBasicTest = @"
 Imports System.Xml
 Imports System.IO
 
@@ -762,7 +637,7 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, Expected).ConfigureAwait(false);
 
             const string cSharpTest2 = @"
 using System.Xml;
@@ -776,7 +651,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest2 = @"
+            const string visualBasicTest2 = @"
 Imports System.Xml
 Imports System.IO
 
@@ -786,9 +661,10 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, Expected).ConfigureAwait(false);
         }
 
+        [TestCategory("Ignore")]
         [DataRow("XmlResolver",   "null")]
         [DataRow("XmlResolver",   SecureResolverText)]
         [DataRow("ProhibitDtd",   "true")]
@@ -802,7 +678,7 @@ End Class
             var vbValue = value.Replace("null", "Nothing");
             vbValue = Regex.Replace(vbValue, @"^\(([^\)]*)\)(.*)", "DirectCast($2, $1)", RegexOptions.Multiline);
 
-            string cSharpTest = $@"
+            var cSharpTest = $@"
 using System.Xml;
 using System.IO;
 
@@ -834,7 +710,7 @@ End Class
 ";
             await VerifyNoWarnings(cSharpTest, visualBasicTest).ConfigureAwait(false);
 
-            string cSharpTest2 = $@"
+            var cSharpTest2 = $@"
 using System.Xml;
 using System.IO;
 
@@ -864,7 +740,7 @@ End Class
 ";
             await VerifyNoWarnings(cSharpTest2, visualBasicTest2).ConfigureAwait(false);
 
-            string cSharpTest3 = $@"
+            var cSharpTest3 = $@"
 using System.Xml;
 using System.IO;
 
@@ -893,6 +769,7 @@ End Class
             await VerifyNoWarnings(cSharpTest3, visualBasicTest3).ConfigureAwait(false);
         }
 
+        [TestCategory("Ignore")]
         [TestMethod]
         public async Task XmlReaderCreateDefaultXmlReaderSettings()
         {
@@ -908,7 +785,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest = @"
+            const string visualBasicTest = @"
 Imports System.Xml
 
 Class Xxe
@@ -931,7 +808,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest2 = @"
+            const string visualBasicTest2 = @"
 Imports System.Xml
 
 Class Xxe
@@ -956,7 +833,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest3 = @"
+            const string visualBasicTest3 = @"
 Imports System.Xml
 Imports System.IO
 
@@ -981,7 +858,7 @@ class Xxe
     }
 }";
 
-            var visualBasicTest4 = @"
+            const string visualBasicTest4 = @"
 Imports System.Xml
 
 Class Xxe
@@ -994,6 +871,7 @@ End Class
             await VerifyNoWarnings(cSharpTest4, visualBasicTest4).ConfigureAwait(false);
         }
 
+        [TestCategory("Ignore")]
         [DataRow("XmlResolver",   "null")]
         [DataRow("XmlResolver",   SecureResolverText)]
         [DataRow("ProhibitDtd",   "true")]
@@ -1007,7 +885,7 @@ End Class
             var vbValue = value.Replace("null", "Nothing");
             vbValue = Regex.Replace(vbValue, @"^\(([^\)]*)\)(.*)", "DirectCast($2, $1)", RegexOptions.Multiline);
 
-            string cSharpTest = $@"
+            var cSharpTest = $@"
 using System.Xml;
 
 class Xxe
@@ -1037,7 +915,7 @@ End Class
 ";
             await VerifyNoWarnings(cSharpTest, visualBasicTest).ConfigureAwait(false);
 
-            string cSharpTest2 = $@"
+            var cSharpTest2 = $@"
 using System.Xml;
 
 class Xxe
@@ -1065,7 +943,7 @@ End Class
 ";
             await VerifyNoWarnings(cSharpTest2, visualBasicTest2).ConfigureAwait(false);
 
-            string cSharpTest3 = $@"
+            var cSharpTest3 = $@"
 using System.Xml;
 
 class Xxe
@@ -1090,92 +968,18 @@ Class Xxe
 End Class
 ";
             await VerifyNoWarnings(cSharpTest3, visualBasicTest3).ConfigureAwait(false);
-
-//            var cSharpTest4 = $@"
-//using System.Xml;
-
-//class Xxe
-//{{
-//    private static XmlReaderSettings Settings = new XmlReaderSettings();
-
-//    public static void parseUpload(string path)
-//    {{
-//#pragma warning disable 618
-//        Settings.{property} = {value};
-//#pragma warning restore 618
-//        XmlReader reader = XmlReader.Create(path, Settings);
-//    }}
-//}}
-//";
-
-//            var visualBasicTest4 = $@"
-//Imports System.Xml
-
-//Class Xxe
-//    Private Shared Settings As XmlReaderSettings = New XmlReaderSettings()
-
-//    Public Shared Sub parseUpload(path As String)
-//#Disable Warning BC40000
-//        Settings.{property} = {vbValue}
-//#Enable Warning BC40000
-//        Dim reader As XmlReader = XmlReader.Create(path, Settings)
-//    End Sub
-//End Class
-//";
-
-//            await VerifyNoWarnings(cSharpTest4, visualBasicTest4).ConfigureAwait(false); // todo: defaults are assumed for member variables
-
-//            var cSharpTest5 = $@"
-//using System.Xml;
-
-//class Xxe
-//{{
-//#pragma warning disable 618
-//    private static XmlReaderSettings Settings = new XmlReaderSettings {{{property} = {value}}};
-//#pragma warning restore 618
-
-//    public static void parseUpload(string path)
-//    {{
-//        XmlReader reader = XmlReader.Create(path, Settings);
-//    }}
-//}}
-//";
-
-//            var visualBasicTest5 = $@"
-//Imports System.Xml
-
-//Class Xxe
-//#Disable Warning BC40000
-//    Private Shared Settings As New XmlReaderSettings With {{.{property} = {vbValue}}}
-//#Enable Warning BC40000
-
-//    Public Shared Sub parseUpload(path As String)
-//        Dim reader As XmlReader = XmlReader.Create(path, Settings)
-//    End Sub
-//End Class
-//";
-
-//            await VerifyNoWarnings(cSharpTest5, visualBasicTest5).ConfigureAwait(false); // todo: defaults are assumed for member variables
         }
 
+        [TestCategory("Detect")]
         [DataRow("ProhibitDtd   = false")]
         [DataRow("DtdProcessing = DtdProcessing.Parse")]
         [DataRow("DtdProcessing = (DtdProcessing)2")]
         [DataTestMethod]
         public async Task XmlReaderCreateDtdProcessingEnabled(string dtdProcessing)
         {
-            var expected = new[]
-            {
-                new DiagnosticResult
-                {
-                    Id       = "SCS0007",
-                    Severity = DiagnosticSeverity.Warning
-                }
-            };
-
             var vbDtdProcessing = Regex.Replace(dtdProcessing, "\\(([^\\)]*)\\)(.*)", "DirectCast($2, $1)");
 
-            string cSharpTest = $@"
+            var cSharpTest = $@"
 using System.Xml;
 
 class Xxe
@@ -1203,9 +1007,9 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest, visualBasicTest, Expected).ConfigureAwait(false);
 
-            string cSharpTest2 = $@"
+            var cSharpTest2 = $@"
 using System.Xml;
 
 public class TestClass
@@ -1231,9 +1035,9 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest2, visualBasicTest2, Expected).ConfigureAwait(false);
 
-            string cSharpTest3 = $@"
+            var cSharpTest3 = $@"
 using System.Xml;
 
 public class TestClass
@@ -1257,99 +1061,7 @@ Class Xxe
     End Sub
 End Class
 ";
-            await VerifyDiagnosticSafeUnsafe(cSharpTest3, visualBasicTest3, expected).ConfigureAwait(false);
+            await VerifyDiagnosticSafeUnsafe(cSharpTest3, visualBasicTest3, Expected).ConfigureAwait(false);
         }
-    }
-
-    [TestClass]
-    public class XxeAnalyzerTaintTest : BaseXxeAnalyzerTest
-    {
-        protected override IEnumerable<DiagnosticAnalyzer> GetDiagnosticAnalyzers(string language)
-        {
-            return new DiagnosticAnalyzer[] { new XxeDiagnosticAnalyzerCSharp(), new XxeDiagnosticAnalyzerVisualBasic(), new TaintAnalyzerCSharp(), new TaintAnalyzerVisualBasic(), };
-        }
-
-        [TestMethod]
-        [Ignore] // todo: taint transfer is not implemented
-        public async Task XmlDocumentLoadTaint()
-        {
-            const string cSharpTest = @"
-using System.Xml;
-
-class Xxe
-{
-    public static void parseUpload(string path)
-    {
-        var xmlDoc = new XmlDocument();
-        var xmlDoc2 = xmlDoc;
-        xmlDoc2.XmlResolver = null;
-        xmlDoc.Load(path);
-    }
-}";
-
-            var visualBasicTest = @"
-Imports System.Xml
-
-Class Xxe
-    Public Shared Sub parseUpload(path As String)
-        Dim xmlDoc As New XmlDocument()
-        Dim xmlDoc2 = xmlDoc;
-        xmlDoc2.XmlResolver = Nothing;
-        xmlDoc.Load(path)
-    End Sub
-End Class
-";
-
-            await VerifyNoWarnings(cSharpTest, visualBasicTest).ConfigureAwait(false);
-        }
-
-        // although the input is untrusted
-        // there is not xxe or path injection
-        [TestMethod]
-        public async Task StringReader()
-        {
-            var cSharpTest = @"
-using System.IO;
-using System.Xml;
-
-class PathTraversal
-{
-    public static void Run(string strText)
-    {
-        using (var reader = XmlReader.Create(new StringReader(strText)))
-        {
-            reader.Read();
-        }
-    }
-}
-";
-
-            var visualBasicTest = @"
-Imports System.IO
-Imports System.Xml
-
-Class PathTraversal
-    Public Shared Sub Run(strText As String)
-        Using reader = XmlReader.Create(New StringReader(strText))
-            reader.Read()
-        End Using
-    End Sub
-End Class
-";
-
-            await VerifyNoWarnings(cSharpTest, visualBasicTest).ConfigureAwait(false);
-        }
-
-        // todo: introduce configuration setting to show questionable findings
-        // add tests where ProhibitDtd, DtdProcessing or entire XmlReaderSettings comes from untrusted source
-        //[DataRow("XmlReader.Create(default(Stream))")]
-        //[DataRow("XmlReader.Create(default(Stream), xmlReaderSettingsInput)")]
-        //[DataRow("XmlReader.Create(default(Stream), xmlReaderSettingsInput, default(string))")]
-        //[DataRow("XmlReader.Create(default(Stream), xmlReaderSettingsInput, default(XmlParserContext))")]
-        //[DataRow("XmlReader.Create(default(TextReader))")]
-        //[DataRow("XmlReader.Create(default(TextReader), xmlReaderSettingsInput)")]
-        //[DataRow("XmlReader.Create(default(TextReader), xmlReaderSettingsInput, default(string))")]
-        //[DataRow("XmlReader.Create(default(TextReader), xmlReaderSettingsInput, default(XmlParserContext))")]
-        //[DataRow("XmlReader.Create(default(TextReader), xmlReaderSettingsInput)")]
     }
 }
